src/requirements.txt
aiogram==3.27.0
asyncpg==0.30.0
aiohttp==3.11.16
openai==1.82.0
python-dotenv==1.1.0
